Add banner item of portfolio
